fastapi-template/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
dnspython==2.4.2
fastapi==0.109.0
idna==3.6
importlib-resources==6.1.1
kombu==5.3.5
limits==3.7.0
packaging==23.2
prompt-toolkit==3.0.43
psutil==5.9.7
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0